ticket:11018: Support CLDR time separator (ICU4C)
This change will treat the ':' character in patterns as a symbol
(instead of as a literal) and if the Locale defines timeSeparator then
use that character instead when formatting and accept both when parsing.

It will also define the '.' as an alternate time separator, that will be
accepted when parsing in lenient mode.

This is one character more strict than before, when in lenient mode also
the '-' was accepted as time separator instead of ':'.

R=rocketman@google.com

Review URL: https://codereview.appspot.com/180610043